Replace DNSimple provider's API v1 endpoints with those from API v2
Rename provider dnsimplev2 to dnsimple, replacing the existing provider
Replace the tests from dnsimple from those for the v2 endpoint
