+++ README.md
@@ -90,13 +90,13 @@ and with Route 53 provider depedencies:
 
     $ lexicon -h
       usage: lexicon [-h] [--version] [--delegated DELEGATED]
-                     {cloudflare,cloudxns,digitalocean,dnsimple,dnsimplev2,dnsmadeeasy,dnspark,dnspod,easydns,luadns,namesilo,nsone,pointhq,rage4,route53,vultr,yandex,zonomi}
+                     {cloudflare,cloudxns,digitalocean,dnsimple,dnsmadeeasy,dnspark,dnspod,easydns,luadns,namesilo,nsone,pointhq,rage4,route53,vultr,yandex,zonomi}
                      ...
 
       Create, Update, Delete, List DNS entries
 
       positional arguments:
-        {cloudflare,cloudxns,digitalocean,dnsimple,dnsimplev2,dnsmadeeasy,dnspark,dnspod,easydns,luadns,namesilo,nsone,pointhq,rage4,route53,vultr,yandex,zonomi}
+        {cloudflare,cloudxns,digitalocean,dnsimple,dnsmadeeasy,dnspark,dnspod,easydns,luadns,namesilo,nsone,pointhq,rage4,route53,vultr,yandex,zonomi}
                               specify the DNS provider to use
           cloudflare          cloudflare provider
           cloudxns            cloudxns provider
