--- tests/fixtures/cassettes/dnsimplev2/IntegrationTests/test_Provider_authenticate.yaml
@@ -1,70 +1,0 @@
-interactions:
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/accounts
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":475,"plan_identifier":"dnsimple-personal","created_at":"2017-03-11T04:50:42Z","updated_at":"2017-03-12T08:06:55Z"}]}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['156']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:53:45 GMT']
-      etag: [W/"72b5baad57ee66be3c25031e6dd757be"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2394']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [088aa3d4-2465-4390-8211-3e6fd61af5b9]
-      x-runtime: ['0.031540']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/domains?name_like=wopr.tech
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":32240,"account_id":475,"registrant_id":null,"name":"wopr.tech","unicode_name":"wopr.tech","state":"hosted","auto_renew":false,"private_whois":false,"expires_on":null,"created_at":"2017-03-12T23:53:34Z","updated_at":"2017-03-12T23:53:34Z"}],"pagination":{"current_page":1,"per_page":30,"total_entries":1,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['378']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:53:45 GMT']
-      etag: [W/"f8acd1dd7c5dc59b6f1af98347e840ee"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2391']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [e3cf6487-2117-4d9e-8e0f-af3c3e7e3304]
-      x-runtime: ['0.029557']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-version: 1
