--- tests/fixtures/cassettes/dnsimplev2/IntegrationTests/test_Provider_when_calling_create_record_for_A_with_valid_name_and_content.yaml
@@ -1,137 +1,0 @@
-interactions:
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/accounts
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":475,"plan_identifier":"dnsimple-personal","created_at":"2017-03-11T04:50:42Z","updated_at":"2017-03-12T08:06:55Z"}]}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['156']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:53:46 GMT']
-      etag: [W/"72b5baad57ee66be3c25031e6dd757be"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2392']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [d47ff7d5-ccbe-4323-a2a0-7ddbdbcfe331]
-      x-runtime: ['0.019431']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/domains?name_like=wopr.tech
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":32240,"account_id":475,"registrant_id":null,"name":"wopr.tech","unicode_name":"wopr.tech","state":"hosted","auto_renew":false,"private_whois":false,"expires_on":null,"created_at":"2017-03-12T23:53:34Z","updated_at":"2017-03-12T23:53:34Z"}],"pagination":{"current_page":1,"per_page":30,"total_entries":1,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['378']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:53:47 GMT']
-      etag: [W/"f8acd1dd7c5dc59b6f1af98347e840ee"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2389']
-      x-ratelimit-reset: ['1489366393']
-      x-request-id: [75f6e769-865f-4988-90e6-f6937855378e]
-      x-runtime: ['0.038086']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/zones/wopr.tech/records?type=A&name=localhost
-  response:
-    body: {string: !!python/unicode '{"data":[],"pagination":{"current_page":1,"per_page":30,"total_entries":0,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['91']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:53:47 GMT']
-      etag: [W/"0605e329ddd741bf0b60c52b68e9846b"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2388']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [9c488a24-767c-4c7b-b564-36de23507b94]
-      x-runtime: ['0.021179']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{"content": "127.0.0.1", "type": "A", "name": "localhost"}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['58']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: POST
-    uri: https://api.sandbox.dnsimple.com/v2475/zones/wopr.tech/records
-  response:
-    body: {string: !!python/unicode '{"data":{"id":311244,"zone_id":"wopr.tech","parent_id":null,"name":"localhost","content":"127.0.0.1","ttl":3600,"priority":null,"type":"A","regions":["global"],"system_record":false,"created_at":"2017-03-12T23:53:47Z","updated_at":"2017-03-12T23:53:47Z"}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:53:47 GMT']
-      etag: [W/"93f0d8a723647a0f245b48cb3b12e716"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2387']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [9e08b448-f2ee-4c81-9f9e-9270c9d9535f]
-      x-runtime: ['0.092807']
-      x-xss-protection: [1; mode=block]
-    status: {code: 201, message: Created}
-version: 1
