--- tests/fixtures/cassettes/dnsimplev2/IntegrationTests/test_Provider_when_calling_list_records_after_setting_ttl.yaml
@@ -1,172 +1,0 @@
-interactions:
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/accounts
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":475,"plan_identifier":"dnsimple-personal","created_at":"2017-03-11T04:50:42Z","updated_at":"2017-03-12T08:06:55Z"}]}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['156']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:07 GMT']
-      etag: [W/"72b5baad57ee66be3c25031e6dd757be"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2383']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [a4ac8e55-c245-4f04-9a68-85885ddb9283]
-      x-runtime: ['0.036366']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/domains?name_like=wopr.tech
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":32240,"account_id":475,"registrant_id":null,"name":"wopr.tech","unicode_name":"wopr.tech","state":"hosted","auto_renew":false,"private_whois":false,"expires_on":null,"created_at":"2017-03-12T23:53:34Z","updated_at":"2017-03-12T23:53:34Z"}],"pagination":{"current_page":1,"per_page":30,"total_entries":1,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['378']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:07 GMT']
-      etag: [W/"f8acd1dd7c5dc59b6f1af98347e840ee"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2350']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [693dfb29-0933-48e2-b0eb-0dc9e54da28e]
-      x-runtime: ['0.042943']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/zones/wopr.tech/records?type=TXT&name=ttl.fqdn
-  response:
-    body: {string: !!python/unicode '{"data":[],"pagination":{"current_page":1,"per_page":30,"total_entries":0,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['91']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:07 GMT']
-      etag: [W/"0605e329ddd741bf0b60c52b68e9846b"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2349']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [52c7342d-2b81-4973-aa58-f8fe5000e089]
-      x-runtime: ['0.037802']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{"content": "ttlshouldbe3600", "type": "TXT", "name":
-      "ttl.fqdn", "ttl": 3600}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['78']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: POST
-    uri: https://api.sandbox.dnsimple.com/v2475/zones/wopr.tech/records
-  response:
-    body: {string: !!python/unicode '{"data":{"id":311253,"zone_id":"wopr.tech","parent_id":null,"name":"ttl.fqdn","content":"ttlshouldbe3600","ttl":3600,"priority":null,"type":"TXT","regions":["global"],"system_record":false,"created_at":"2017-03-12T23:54:08Z","updated_at":"2017-03-12T23:54:08Z"}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:08 GMT']
-      etag: [W/"04936e2834afc1d6d5a03b8ee4244517"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2348']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [28ca7dd0-e7be-4d5a-a925-875cc73e423b]
-      x-runtime: ['0.130566']
-      x-xss-protection: [1; mode=block]
-    status: {code: 201, message: Created}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/zones/wopr.tech/records?type=TXT&name=ttl.fqdn
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":311253,"zone_id":"wopr.tech","parent_id":null,"name":"ttl.fqdn","content":"ttlshouldbe3600","ttl":3600,"priority":null,"type":"TXT","regions":["global"],"system_record":false,"created_at":"2017-03-12T23:54:08Z","updated_at":"2017-03-12T23:54:08Z"}],"pagination":{"current_page":1,"per_page":30,"total_entries":1,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['344']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:08 GMT']
-      etag: [W/"d360ee74e14be27caea91b0d62e590c2"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2347']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [b35ec6c6-3911-4daa-8858-5223beb8265b]
-      x-runtime: ['0.049417']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-version: 1
