--- tests/fixtures/cassettes/dnsimplev2/IntegrationTests/test_Provider_when_calling_list_records_with_full_name_filter_should_return_record.yaml
@@ -1,171 +1,0 @@
-interactions:
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/accounts
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":475,"plan_identifier":"dnsimple-personal","created_at":"2017-03-11T04:50:42Z","updated_at":"2017-03-12T08:06:55Z"}]}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['156']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:11 GMT']
-      etag: [W/"72b5baad57ee66be3c25031e6dd757be"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2381']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [e1114d7e-99c6-4b27-be2e-b2cc208ec5e6]
-      x-runtime: ['0.026906']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/domains?name_like=wopr.tech
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":32240,"account_id":475,"registrant_id":null,"name":"wopr.tech","unicode_name":"wopr.tech","state":"hosted","auto_renew":false,"private_whois":false,"expires_on":null,"created_at":"2017-03-12T23:53:34Z","updated_at":"2017-03-12T23:53:34Z"}],"pagination":{"current_page":1,"per_page":30,"total_entries":1,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['378']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:11 GMT']
-      etag: [W/"f8acd1dd7c5dc59b6f1af98347e840ee"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2342']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [15e30a9e-fc6c-4e43-9c7a-8dad5e43dd0b]
-      x-runtime: ['0.030959']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/zones/wopr.tech/records?type=TXT&name=random.fulltest
-  response:
-    body: {string: !!python/unicode '{"data":[],"pagination":{"current_page":1,"per_page":30,"total_entries":0,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['91']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:12 GMT']
-      etag: [W/"0605e329ddd741bf0b60c52b68e9846b"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2341']
-      x-ratelimit-reset: ['1489366393']
-      x-request-id: [675d9d9b-dcf3-488a-995a-bcad00b991a5]
-      x-runtime: ['0.043001']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: !!python/unicode '{"content": "challengetoken", "type": "TXT", "name": "random.fulltest"}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['71']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: POST
-    uri: https://api.sandbox.dnsimple.com/v2475/zones/wopr.tech/records
-  response:
-    body: {string: !!python/unicode '{"data":{"id":311255,"zone_id":"wopr.tech","parent_id":null,"name":"random.fulltest","content":"challengetoken","ttl":3600,"priority":null,"type":"TXT","regions":["global"],"system_record":false,"created_at":"2017-03-12T23:54:12Z","updated_at":"2017-03-12T23:54:12Z"}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:12 GMT']
-      etag: [W/"d7a18c9c0692179eda1d714b4ac9fea2"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2340']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [895a574d-b8e0-40cc-84d1-7014ad4a12d6]
-      x-runtime: ['0.080676']
-      x-xss-protection: [1; mode=block]
-    status: {code: 201, message: Created}
-- request:
-    body: !!python/unicode '{}'
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['2']
-      Content-Type: [application/json]
-      User-Agent: [python-requests/2.13.0]
-    method: GET
-    uri: https://api.sandbox.dnsimple.com/v2/475/zones/wopr.tech/records?type=TXT&name=random.fulltest
-  response:
-    body: {string: !!python/unicode '{"data":[{"id":311255,"zone_id":"wopr.tech","parent_id":null,"name":"random.fulltest","content":"challengetoken","ttl":3600,"priority":null,"type":"TXT","regions":["global"],"system_record":false,"created_at":"2017-03-12T23:54:12Z","updated_at":"2017-03-12T23:54:12Z"}],"pagination":{"current_page":1,"per_page":30,"total_entries":1,"total_pages":1}}'}
-    headers:
-      cache-control: ['max-age=0, private, must-revalidate']
-      connection: [keep-alive]
-      content-length: ['350']
-      content-type: [application/json; charset=utf-8]
-      date: ['Sun, 12 Mar 2017 23:54:12 GMT']
-      etag: [W/"8d725d892f9bbb0b3ff4d8aeffa4d0b8"]
-      server: [nginx]
-      strict-transport-security: [max-age=31536000]
-      transfer-encoding: [chunked]
-      x-content-type-options: [nosniff]
-      x-download-options: [noopen]
-      x-frame-options: [DENY]
-      x-permitted-cross-domain-policies: [none]
-      x-ratelimit-limit: ['2400']
-      x-ratelimit-remaining: ['2339']
-      x-ratelimit-reset: ['1489366392']
-      x-request-id: [52f7dcf0-0ff3-4da1-b535-e641c2b0cd8d]
-      x-runtime: ['0.043966']
-      x-xss-protection: [1; mode=block]
-    status: {code: 200, message: OK}
-version: 1
